backend/requirements.txt
Django
djangorestframework
pymongo
PyMuPDF
crewai
crewai-tools
google-generativeai
requests
python-dotenv
